Business Jet Ticket Purchasing Program
This program was mostly completed as an individual project. The goal was to create a mock-up version of an airplane ticket purchasing website. The program presents three different sections of the plane, with each section holding a different seating arrangement. The user is able to select which section they want to sit in, input how many seats they would like to purchase, then select which seats they would like to purchase based on a virtual representation of the seats. The user is also able to change their minds on seats they have already purchased.