Hardware Store via Vectors
Pair project that creates a store selling miscellaneous items. The inventory displays certain information on each item, such as item stock, selling price, back orders, and the mark up percentage. The user is able to see if a certain item is in stock, purchase an item, check the price of an item, display the inventory, and return an item.